backend/requirements.txt
# ===================================
# Core Web Framework
# ===================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# ===================================
# Database & ORM
# ===================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# ===================================
# Authentication & Security
# ===================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# ===================================
# API Documentation
# ===================================
python-dateutil==2.8.2
email-validator==2.1.0

# ===================================
# Voice-to-Text AI Services
# ===================================
openai==1.3.5
# google-cloud-speech==2.21.0  # uncomment if needed
SpeechRecognition==3.10.0

# ===================================
# File Handling
# ===================================
pydub==0.25.1
python-magic==0.4.27

# ===================================
# Caching & Performance
# ===================================
redis==5.0.1
hiredis==2.2.3

# ===================================
# Environment & Config
# ===================================
python-dotenv==1.0.0

# ===================================
# HTTP Client
# ===================================
httpx==0.25.1
aiofiles==23.2.1

# ===================================
# Testing (Development)
# ===================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# ===================================
# Code Quality (Development)
# ===================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# ===================================
# Monitoring & Logging
# ===================================
python-json-logger==2.0.7

# ===================================
# Rate Limiting
# ===================================
slowapi==0.1.9
